requirements.txt
pandas==2.3.3
requests==2.32.5
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
gspread==6.2.1
gspread-dataframe==4.0.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
schedule==1.2.2